requirements.txt
PyYAML
Flask
python-dotenv
google-genai